Improve scalability of tests
